- Move pds3file_rules.py into pds3file/__init__.py - Move pds4file_rules.py into pds4file/__init__.py - Remove unused comments in pdsfile/__init__.py
